Remove BuildImage method from client and move jsonmessage.go to client from docker.go
Removing the BuildImage method allows to remove all dependencies on the `archive` package from the docker library dependcy which helps with crosscompiling.

Moving jsonmesssge.go makes it possible to do all changes directly in the client and not have to change docker itself. Makes updating it easier.
Remove all terminal dependencies as this is not an expected interaction pattern.